staging: Grouping all functions related to installation
Move to staging.c all functions related to installation or removal of files
in the system. Reorganize the staging_install_all_files() too.
